updated workflow for internal version checking
includes output files with the final L1 data for Canada & CONUS. However, @billspat I will be creating a separate script just to simply produce this CSV (canada.conus.splits_L1.csv) so that the internal checking is not included in the final repo. It will be a copy of the versioncheck script content that creates canada.conus.splits_L1.csv. Temp file is some code that also works for a part of the versioncheck script (as an alternate approach).