updating to ensure that Northwestern Crow is included (but changed to American Crow)
This keeps 3 species in the BBS observation data (with their 3 original AOUs), but identifies them as a single species (American Crow) and its AOU.combo is assigned to all species (34880). New output for species list increases the number of species by 2, even though they are the same species (because this will help with merging into the BBS obs data). Summary info is provided at the end to track number of unique species and unique species.combo.